shopping_list/backend/requirements.txt
alembic==1.8.1
asyncpg==0.27.0
greenlet==2.0.1
grpcio==1.51.1
grpcio-tools==1.51.1
importlib-metadata==5.1.0
importlib-resources==5.10.1
Mako==1.2.4
MarkupSafe==2.1.1
protobuf==4.21.11
psycopg2-binary==2.9.5
pydantic==1.10.2
SQLAlchemy==1.4.45
typing-extensions==4.4.0
zipp==3.11.0
